+++ pages/_app.js
@@ -6,6 +6,7 @@ import CssBaseline from '@mui/material/CssBaseline';
 import { CacheProvider } from '@emotion/react';
 import theme from '../src/theme';
 import createEmotionCache from '../src/createEmotionCache';
+import '../styles.css'
 
 const clientSideEmotionCache = createEmotionCache();
